backend/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
nest-asyncio>=1.5.6

# Web framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Request handling
requests>=2.31.0
aiofiles>=23.2.1

# Azure services
azure-core>=1.29.5
azure-search-documents>=11.4.0
azure-ai-formrecognizer>=3.3.0
azure-storage-blob>=12.19.0

# Image processing
Pillow>=10.1.0

# LlamaIndex components
llama-index-core>=0.10.0
llama-index-embeddings-azure-openai>=0.1.5
llama-index-llms-azure-openai>=0.1.5
llama-index-multi-modal-llms-azure-openai>=0.1.3
llama-index-vector-stores-azureaisearch>=0.1.3

# Monitoring and logging
python-json-logger>=2.0.7

# Rate limiting and caching
redis>=5.0.1
flask-limiter>=3.5.0
flask-caching>=2.1.0

# Security
cryptography>=41.0.7

# Health checks
psutil>=5.9.6